refactor: move class StringAddCalculator
